Block group descriptors checksum support being added.

+++ ext4_crc32.c
@@ -42,9 +42,7 @@
  * CRC32 code derived from work by Gary S. Brown.
  */
 
-#include <stdint.h>
-#include <stddef.h>
-#include <sys/cdefs.h>
+#include "types/ext4_basic.h"
 
 const uint32_t crc32_tab[] = {
 	0x00000000, 0x77073096, 0xee0e612c, 0x990951ba, 0x076dc419, 0x706af48f,
@@ -200,10 +198,9 @@ static const uint32_t crc32Table[256] = {
 };
 
 uint32_t
-ext4_crc32c(uint32_t crc, const void *buf, size_t size)
+ext4_crc32c(uint32_t crc, const void *buf, int size)
 {
 	const uint8_t *p = buf;
-
 
 	while (size--)
 		crc = crc32Table[(crc ^ *p++) & 0xff] ^ (crc >> 8);
